修改类的属性

+++ guest/src/main/java/com/guest/pojo/po/Menu.java
@@ -14,6 +14,6 @@ public class Menu extends Model<Menu> {
     @TableId(value = "dish_id")
     int dish_id;//菜编号
     String dish_name;//菜名
-    float dish_price;//菜价
+    double dish_price;//菜价
 
 }
